Ajustes de banco e diagramas atualizados
