+++ .gitignore
@@ -1,5 +1,4 @@
 env
-__*
 __LibManager
 __Py*
 __pycache__/
