+++ models/__init__.py
@@ -1,0 +1,7 @@
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
+
+def init_app(app):
+    db.init_app(app)
+    return db
